update brats with safe casting

+++ yucca/pipeline/task_conversion/Task012_BraTS21.py
@@ -53,7 +53,7 @@ def convert(path: str, subdir: str = "brats21/training_data"):
         labelarr = label.get_fdata()
         labelarr[labelarr == 4.0] = 3.0
         assert np.all(np.isin(np.unique(labelarr), np.array([0, 1, 2, 3])))
-        labelnew = nib.Nifti1Image(labelarr, label.affine, label.header)
+        labelnew = nib.Nifti1Image(labelarr, label.affine, label.header, dtype=np.float32)
         nib.save(labelnew, dst_label_path)
 
         shutil.copy2(src_image_file_path1, dst_image_file_path1)
@@ -76,7 +76,7 @@ def convert(path: str, subdir: str = "brats21/training_data"):
         labelarr = label.get_fdata()
         labelarr[labelarr == 4.0] = 3.0
         assert np.all(np.isin(np.unique(labelarr), np.array([0, 1, 2, 3])))
-        labelnew = nib.Nifti1Image(labelarr, label.affine, label.header)
+        labelnew = nib.Nifti1Image(labelarr, label.affine, label.header, dtype=np.float32)
         nib.save(labelnew, dst_label_path)
 
         shutil.copy2(src_image_file_path1, dst_image_file_path1)
